This commit contains the update of spot-ml from Spark 1.6 to Spark 2.1. This new version introduces SparkSession as the new point of entry for SparkContext, SQLContext and HiveContext used in Spark 1.x
Changed "tdur" variable from float to double because according to the spot-ml documentation this is the correct data type.

log4j.logger was changed to report ERROR instead of WARN + ERROR

Unit tests were modified to be able to run a local SparkSession instead of the old SparkContext

yarn-client was changed to yarn

Locals folders were changed from Scala 2.10 to 2.11

minExecutors in ml_test.sh was changed to the default 0 to avoid warnings while running the code complaining that initial executors = 1 is not < that minExecutors = 1

+++ spot-ml/src/main/scala/org/apache/spot/netflow/FlowSchema.scala
@@ -49,7 +49,7 @@ object FlowSchema {
   val SecondField = StructField(Second, IntegerType, nullable = true)
 
   val Duration = "tdur"
-  val DurationField = StructField(Duration, FloatType, nullable = true)
+  val DurationField = StructField(Duration, DoubleType, nullable = true)
 
   val SourceIP = "sip"
   val SourceIPField = StructField(SourceIP, StringType, nullable = true)
